requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cassandra-driver==3.28.0
click==8.1.7
colorama==0.4.6
DateTime==5.2
Deprecated==1.2.14
dnspython==2.4.2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
flask-swagger==0.2.14
geomet==0.2.1.post1
git-filter-repo==2.38.0
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mdurl==0.1.2
mysql-connector-python==8.1.0
mysqlclient==2.2.0
ordered-set==4.1.0
packaging==23.1
protobuf==4.21.12
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.5.0
pyotp==2.9.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rich==13.5.2
schedule==1.2.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==3.0.1
zipp==3.17.0
zope.interface==6.0
